requirements.txt
streamlit
openai
python-dotenv
pydantic
pydantic-settings
dashscope
pandas
folium
streamlit-folium
reportlab